+++ src/layout/layoutcomponents.js
@@ -14,5 +14,6 @@ import Matrix from "./matrix"
 import NBody from "./nbody"
 import Radial from "./radial"
 import Tree from "./tree"
+import Force from "./force"
 
-export { Animation, Attraction, BoundingBox, Center, Cluster, Collision, D3Adapter, Fan, LayoutComponent, Grid, Hierarchy, Link, Matrix, NBody, Radial, Tree }
+export { Animation, Attraction, BoundingBox, Center, Cluster, Collision, D3Adapter, Fan, LayoutComponent, Grid, Hierarchy, Link, Matrix, NBody, Radial, Tree, Force }